% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{getAnalytes,ProteomicsNavigator-method}
\alias{getAnalytes,ProteomicsNavigator-method}
\alias{getAnalytes}
\title{the sorted, combined, unique list of analytes, rownames of the 'avg' matrices}
\usage{
\S4method{getAnalytes}{ProteomicsNavigator}(obj)
}
\description{
the sorted, combined, unique list of analytes, rownames of the 'avg' matrices
}
