% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{getDataSet,ProteomicsNavigator-method}
\alias{getDataSet,ProteomicsNavigator-method}
\alias{getDataSet}
\title{return 2 matrices: avg & sd}
\usage{
\S4method{getDataSet}{ProteomicsNavigator}(obj, dataSetName)
}
\arguments{
\item{obj}{a ProteomicsNavigator instance}

\item{dataSetName}{a character string}
}
\description{
return 2 matrices: avg & sd
}
