% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{getDataSetNames,ProteomicsNavigator-method}
\alias{getDataSetNames,ProteomicsNavigator-method}
\alias{getDataSetNames}
\title{discern the experiment names, their constant values, and the experimental manipulations}
\usage{
\S4method{getDataSetNames}{ProteomicsNavigator}(obj)
}
\description{
discern the experiment names, their constant values, and the experimental manipulations
}
