% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{getExperimentMetadata,ProteomicsNavigator-method}
\alias{getExperimentMetadata,ProteomicsNavigator-method}
\alias{getExperimentMetadata}
\title{get metadata for the specified experiment (aka dataset)}
\usage{
\S4method{getExperimentMetadata}{ProteomicsNavigator}(obj, dataSetName)
}
\description{
get metadata for the specified experiment (aka dataset)
}
