% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{getPlotFriendlyDataFrames,ProteomicsNavigator-method}
\alias{getPlotFriendlyDataFrames,ProteomicsNavigator-method}
\alias{getPlotFriendlyDataFrames}
\title{a list of dataframes, each ready for plotting}
\usage{
\S4method{getPlotFriendlyDataFrames}{ProteomicsNavigator}(obj, analytes, experiments)
}
\description{
a list of dataframes, each ready for plotting
}
