% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\name{loadDataSets,ProteomicsNavigator-method}
\alias{loadDataSets,ProteomicsNavigator-method}
\alias{loadDataSets}
\title{discern the experiment names, their constant values, and the experimental manipulations}
\usage{
\S4method{loadDataSets}{ProteomicsNavigator}(obj)
}
\description{
discern the experiment names, their constant values, and the experimental manipulations
}
