% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsNavigator.R
\docType{class}
\name{ProteomicsNavigator-class}
\alias{ProteomicsNavigator-class}
\alias{.ProteomicsNavigator}
\alias{ProteomicsNavigator}
\title{ProteomicsNavigator}
\usage{
ProteomicsNavigator(datasetFilePaths)
}
\arguments{
\item{tabDelimitedFilename}{character}
}
\value{
An object of the ProteomicsNavigator class
}
\description{
ProteomicsNavigator

Create a ProteomicsNavigator connection
}
